pyproject.toml
[project]
name = "evelynn-discordbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "cachetools>=6.1.0",
    "py-cord>=2.6.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "sqlmodel>=0.0.24",
]
